% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{assign_table_function}
\alias{assign_table_function}
\title{Create one table function}
\usage{
assign_table_function(table_name, con_id, env = parent.frame())
}
\arguments{
\item{table_name}{Name of the table}

\item{con_id}{A short string describing a globally cached connection pool}

\item{env}{Environment in which to create the table accessors, such
as the global environment or a package namespace.}
}
\description{
Create one table function
}
