% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{dbc_add_connection}
\alias{dbc_add_connection}
\title{Add a connection or pool to the global options}
\usage{
dbc_add_connection(con, con_id)
}
\arguments{
\item{con}{Connection or pool object}

\item{con_id}{A short string describing a globally cached connection pool}
}
\description{
dbcooper maintains a named list of connections or
(recommended) pools. This internal function adds
one.
}
