% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{dbc_bigquery}
\alias{dbc_bigquery}
\title{Wrap a BigQuery dataset with accessor functions}
\usage{
dbc_bigquery(
  prefix,
  dataset,
  billing = dbc_param("BIGQUERY_BILLING_PROJECT"),
  project = "bigquery-public-data",
  env = parent.frame(),
  ...
)
}
\description{
Wrap a BigQuery dataset with accessor functions
}
