% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{dbc_execute}
\alias{dbc_execute}
\title{Execute a query on a SQL database}
\usage{
dbc_execute(query, con_id)
}
\arguments{
\item{query}{Either a SQL query as a string, a file containing a SQL
query, or a YAML file with a \code{sql} parameter.}

\item{con_id}{A short string describing a globally cached connection pool}
}
\description{
Execute a query on a SQL database
}
