% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{dbc_get_connection}
\alias{dbc_get_connection}
\title{Retrieve a connection or pool from the global options}
\usage{
dbc_get_connection(con_id)
}
\arguments{
\item{con_id}{A short string describing a globally cached connection pool}
}
\description{
dbcooper maintains a named list of connections or
(recommended) pools. This internal function retrieves
one.
}
