% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbc_param}
\alias{dbc_param}
\title{Require a parameter by name}
\usage{
dbc_param(param_name)
}
\arguments{
\item{param_name}{Parameter to look up with \code{\link[base]{Sys.getenv}}.}
}
\description{
Throws an error if the parameter is not set.
}
