% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{dbc_query}
\alias{dbc_query}
\title{Run a query on a SQL database and get a remote table back}
\usage{
dbc_query(query, con_id)
}
\arguments{
\item{query}{Either a SQL query as a string, a file containing a SQL
query, or a YAML file with a \code{sql} parameter.}

\item{con_id}{A short string describing a globally cached connection pool}
}
\value{
A \code{tbl_sql} with the remote results
}
\description{
Run a query on a SQL database and get a remote table back
}
