% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{dbc_table}
\alias{dbc_table}
\title{Access a single table}
\usage{
dbc_table(table_name = NULL, con_id)
}
\arguments{
\item{table_name}{Name of the table in the database. If no table is
provided, returns a list of tables as a character vector.}

\item{con_id}{A short string describing a globally cached connection pool}
}
\description{
Access a single table
}
