% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{query_from_str}
\alias{query_from_str}
\title{Given a string, turn it into a SQL query}
\usage{
query_from_str(query)
}
\arguments{
\item{query}{A string or filename}
}
\description{
Internal function to pull a query from a string. If the string is in a
YAML or plain text file, read it
}
