% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{rmdb_tbl}
\alias{rmdb_tbl}
\title{Access a table by name}
\usage{
rmdb_tbl(table_name)
}
\arguments{
\item{table_name}{a table name to send to the database}
}
\description{
Access a table by name
}
