% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{Set an option by name}
\usage{
set_option(option_name, value)
}
\arguments{
\item{option_name}{Name}

\item{value}{Value}
}
\description{
Utility. This is a bit difficult to do with the options() function.
}
